Develop

+++ ManageIt/src/ManageIt.Application/UseCases/Products/Delete/DeleteProductUseCase.cs
@@ -8,15 +8,24 @@ namespace ManageIt.Application.UseCases.Products.Delete
     internal class DeleteProductUseCase : IDeleteProductUseCase
     {
         private readonly IProductWriteOnlyRepository _repository;
+        private readonly IProductReadOnlyRepository _readOnlyrepository;
         private readonly IUnitOfWork _unitOfWork;
-        public DeleteProductUseCase(IProductWriteOnlyRepository repository, IUnitOfWork unitOfWork)
+        public DeleteProductUseCase(IProductWriteOnlyRepository repository, IProductReadOnlyRepository readOnlyrepository, IUnitOfWork unitOfWork)
         {
             _repository = repository;
+            _readOnlyrepository = readOnlyrepository;
             _unitOfWork = unitOfWork;
         }
 
-        public async Task Execute(Guid id)
+        public async Task Execute(Guid id, Guid companyId)
         {
+            var getById = await _readOnlyrepository.GetById(id);
+
+            if (getById.CompanyId != companyId)
+            {
+                throw new NotFoundException(ResourceErrorMessages.COLLABORATOR_NOT_FOUND);
+            }
+
             var result = await _repository.Delete(id);
 
             if (result is false)
